feat: Scalar serialization and pyzx interop (#33)
Adds an encoding definition for serialising scalars into the pyzx
format.

- Removed the `ignore_scalar` option in the encoder/decorder. We can now
deserialise the full thing.
- Added better error support during serialisation.
- Exposed the scalar from python too. Uses `pyzx.Scalar` for the
user-facing bindings. We just serialize-deserialize between the quizx
and pyzx implementation.

The serialisation incurs a precision loss due to the different encodings
between pyzx and quizx. We should look into this in the future to
improve the situation.