chore: Relax python version req. to include 3.9 (#65)
Title, basically. The current set of dependencies allow for relaxing the
Python dependency in the `pyproject.toml` down to `>=3.9`. I was hoping
to go all the way down to `>=3.7`, since [that is what pyzx
requires](https://github.com/zxcalc/pyzx/blob/master/setup.py#L36), but
this was not possible without downgrading other dependencies.