chore: Cleanup non-standard pyproject location (#99)
Having `pyproject.toml` outside the package directory is non-standard,
and caused issues with `release-please` not updating the package
version.

This PR moves the `pyproject.toml` to the root of the python package,
`pybindings/`, and adds a `uv` workspace at `/` instead.

Also, fixes the package version that didn't get bumped after #85.