refactor: simplified scalar implementation (#109)
Introduces a new `FScalar` class replacing the existing `Scalar<T>`
class. The new class is easier to understand and covers most relevant
use-cases without the disadvantages of the old `Scalar<T>` struct,
namely:
* `Scalar<T>` code is hard to understand, due to a fairly complex
relationship with the `Coeffs` trait bound on `T` and the ability to
multiply hetrogeneous scalar types (necessitating computing LCMs etc)
* The complexity of multiplication is quadratic in the precision of
scalars. For instance, scalars arising from Clifford+$Z[\pi/128]$
circuits require 128*128 floating point multiplications for _each_
multiplication. The extra precision is not really worth it IMO.

The new scalar consists of 4 floats corresponding to the coefficients $a
+ b \omega + c \omega^2 + d \omega^3$ for $\omega := e^{i \pi/4}$. This
allows exact representation for Clifford+T scalars, i.e. numbers in
$\mathbb Z[\sqrt{2}, i]$, and approximate representation of arbitrary
complex numbers.

The exact representation for Clifford+T scalars uses the fact that the
coefficients $a, b, c, d$ are always rational numbers of the form
$z/2^k$ for integers $z, k$. Hence, as long as $z$ and $k$ fit within
the significand and exponent of an `f64`, respectively, these numbers
can be exactly represented by a float.

Arbitrary complex numbers can be approximated as an `FScalar` by noting
that $\omega^2 = i$, so setting $b = d = 0$, we get $a + c i$, which can
be readily converted to/from a `Complex<f64>` from the Rust standard
library.