ci: Setup crate releases (#120)
Sets up `release-plz` to auto-generate release PRs with changelog and
crate version updates (similar to the python ones).
Once one of those PRs gets merged, it will automatically generate a new
github release for the crate and upload it to crates.io.

I also added a workflow that runs `cargo-semver-checks`, so it detects
when the changes in a PR break semver compatibility and we should mark
it as such (by adding a `!` in the PR title). These `!`s will be
detected by `release-plz` when choosing the next crate version.

I haven't added a crates io api token yet. Tokens can be given
permissions to publish new crates and to publish crate-updates
separatedly. To avoid any surprises, I prefer to only give automation
the latter, so I'll do the first release `0.1.0` manually instead.