feat: added method to pack VecGraph vertices, removing holes (#126)
`VecGraph` implements constant-time deletion by leaving "holes" whenever
a vertex is deleted. This means simplifications that remove most of the
nodes from a graph end up eventually spending the majority of their time
skipping over holes. This adds a method `GraphLike::pack` which can be
called periodically to remove all holes, using O(n) time and extra
space. This method takes a flag `force` which tells it whether to always
do packing (`force = true`) or to only pack if the proportion of holes
is over a fixed threshhold (currently set to 10%).