feat!: Added dyadic rational type and new scalars based on this type (#158)
Introduces two new structs `Dyadic` for dyadic rationals and `Scalar4`
which consists for 4 dydadic rational coefficients for powers of $\omega
= e^{i \pi/4}$.

Dyadic rationals store a denominator (a.k.a. mantissa) and a signed
power of 2. They are essentially floating point numbers, with a couple
of exceptions:
- they have a full 32 bits for representing very large/small exponents,
which are needed for renormalising large circuits. `f64` only has 11
bits, so could overflow if a circuit has more than ~2000 gates
- dyadics have an `approx` flag that gets set to true if it ever does
any rounding. If the flag is false after a calculation, this represents
the exact value of a calculation.

For the second point, `Scalar4` also has an `approx()` method. If this
is still false after a calculation, the result is a geniune exact
amplitude in the ring $\mathbb D[\omega]$.

BREAKING CHANGE:

- Graphs and related objects now store a `Scalar4`, which contains 4
`Dyadic` coefficients, rather than 4 `f64` coefficients like `FScalar`
did. Methods that expect these to be floats should use `try_from` /
`try_into` methods to convert to floats.
- `FScalar` has been deprecated. I'm keeping it around for now because
native floating point arithmetic has slightly better performance than
the equivalent dyadic functions. We might want this as an option for
tensor contraction.