feat!: Python bindings for rankwidth functions (#168)
Added Python bindings for rankwidth functions, and some general tidying
of the Python bindings.

BREAKING CHANGE:

Renamed all of the Python wrapper classes to PyXXXX, with annotation
`#[pyclass(name = "XXXX")]`. This should be less confusing to new
developers not to have classes with identical names in the `quizx` and
`quizx_pybindings` crates.