fix!: added missing methods to VecGraph bindings (#170)
PyVecGraph now has implementations for all the methods in BaseGraph,
although some give not implemented exceptions or warnings.
BaseGraph.copy works again as a way of constructing a rust graph from a
python one.

BREAKING CHANGE:

Signatures of some of the methods in the python bindings have changed
for consistency with the PyZX methods. This shouldn't affect any
external code.